🧪 test(test/loc): test loc done
1. test loc done with no template version but slow perf happens again no matter what compiler i use!!

+++ dataproxy/src/LidarDataProxy.cpp
@@ -46,7 +46,7 @@ void LidarDataProxy::visPCHandler()
         case VisType::Aligned: {
             // trans
             pc_t::Ptr aligned(new pc_t());
-            pcl::transformPointCloud(*mAlignedKF.pc, *aligned, mAlignedKF.pose.matrix());
+            pcl::transformPointCloud(*mAlignedKF.pc, *aligned, mAlignedKF.pose.matrix().cast<float>());
             
             sensor_msgs::PointCloud2 rospc;
             pcl::toROSMsg(*aligned, rospc);
